the beginning of the redesign
