+++ src/bible/verse.rs
@@ -1,0 +1,5 @@
+
+pub struct Verse {
+    number: u32,
+    contents: String,
+}
